==> 06-combine-two-mechanisms.ipynb <==
## 06. Merge two mechanisms and get a summary table for each definition.

import pandas as pd
import merge_data as md

VIC_DATA_PATH = "s3://social-research-cheating/summary-tables/emp-net/vic.parquet"
SIM_OBS_DATA_PATH = "s3://social-research-cheating/summary-tables/emp-net/simple_obs.parquet"
STR_OBS_DATA_PATH = "s3://social-research-cheating/summary-tables/emp-net/strict_obs.parquet"

vic_data = md.get_transitions(VIC_DATA_PATH, 7)
vic_data.registerTempTable("vic_data")
vic_data.show(5)
vic_df = vic_data.toPandas()

## 1) Merge two tables and get a summary table for each definition.

obs_data = md.get_transitions(SIM_OBS_DATA_PATH, 7)
obs_data.registerTempTable("obs_data")
obs_data.show(5)
obs_df = obs_data.toPandas()

frequency_table = md.merge_tables(vic_df, obs_df, 0)

print(frequency_table.head())
print(frequency_table.tail())
frequency_table.to_csv('emp_data.csv', index=False)

obs_data = md.get_transitions(STR_OBS_DATA_PATH, 7)
obs_data.registerTempTable("obs_data")
obs_data.show(5)
obs_df = obs_data.toPandas()

frequency_table = md.merge_tables(vic_df, obs_df, 1)

print(frequency_table.head())
print(frequency_table.tail())
frequency_table.to_csv('emp_data.csv', index=False)

## 2) Put the results into csv files.

for i in range(1, 6):
    md.put_summary_table_in_csv_file(i, 0)
    
simple_results = md.create_merged_csv_file("emp_data.csv", "rand_data_1.csv", 5)
simple_results.to_csv("RO_RE.csv", index=False)

print(simple_results.head())
print(simple_results.tail())

for i in range(1, 6):
    md.put_summary_table_in_csv_file(i, 1)

strict_results = md.create_merged_csv_file("emp_data.csv", "rand_data_1.csv", 5)
strict_results.to_csv("SO_SE.csv", index=False)

print(strict_results.head())
print(strict_results.tail())
